requirements.txt
pandas
sqlalchemy
spotipy
python-dotenv
tqdm
psycopg2-binary
